Apply suggestions from code review

+++ src/diffusers/pipelines/versatile_diffusion/modeling_text_unet.py
@@ -394,7 +394,7 @@ class UNetFlatConditionModel(ModelMixin, ConfigMixin):
                 dtype = torch.float32 if is_mps else torch.float64
             else:
                 dtype = torch.int32 if is_mps else torch.int64
-            timesteps = torch.tensor([timesteps], dtype=torch.int32, device=sample.device)
+            timesteps = torch.tensor([timesteps], dtype=dtype, device=sample.device)
         elif len(timesteps.shape) == 0:
             timesteps = timesteps[None].to(sample.device)
